Update auto_patcher
Hey Mateor, I couldn't test this but provided to a user on xda who was getting unknown rom version errors. They are running PACman and their build.prop differs slightly to mine as it is capital PAC. I hope I used the correct syntax to add an OR clause to check for both "pac" and "PAC" which I think may fix the error. Cheers, Aussie.